requirements.txt
python==3.8.19
numpy==1.24.4
opencv-python==4.10.0.84
pandas==2.0.3
pillow==10.4.0
scikit-learn==1.3.2
scipy==1.10.1
timm==0.5.4
torch==2.4.0
torchstat==0.0.7
torchvision==0.19.0
tqdm==4.66.4